LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/layer-keyframes-override-expected.txt


FAIL @keyframes layered overrides unlayered assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
FAIL @keyframes override between layers assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
FAIL @keyframes override update with appended sheet 1 assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
FAIL @keyframes override update with appended sheet 2 assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
